+++ src/Components/Projects/Projects.tsx
@@ -1,0 +1,11 @@
+import "./projects.css";
+
+const Projects = () => {
+  return (
+    <section id="projects" style={{marginTop: 150}}>
+      <h1 className="fs-1 text-end header mb-5">Projects</h1>
+    </section>
+  );
+}
+ 
+export default Projects;