+++ src/Components/index.ts
@@ -1,3 +1,4 @@
 export {default as NavLinks} from "./NavLinks/NavLinks";
 export {default as AboutMe} from "./AboutMe/AboutMe";
 export {default as Skills} from "./Skills/Skills";
+export {default as Projects} from "./Projects/Projects";